Bug: Ne supprimer le fichier compresser de l'export de vue que s'il existe.
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@13601 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/system/classes/CViewSender.class.php
@@ -223,7 +223,10 @@ class CViewSender extends CMbObject {
     }
     
     unlink($this->_file);
-    unlink($this->_file_compressed);
+    
+    if (file_exists($this->_file_compressed)) {
+      unlink($this->_file_compressed);
+    }
   }
   
   /**
